Updated to version 1.4.4 from http://www.the-ebook.org/forum/viewtopic.php?p=1278819#1278819.

+++ src/com/harasoft/relaunch/DropBoxActivity.java
@@ -4,42 +4,21 @@ import android.app.Activity;
 import android.app.ActivityManager;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.content.pm.PackageManager;
-import android.content.res.Configuration;
-import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.TextView;
-import android.widget.Toast;
 import com.dropbox.client2.DropboxAPI;
-import com.dropbox.client2.android.AndroidAuthSession;
-import com.dropbox.client2.android.AuthActivity;
-import com.dropbox.client2.exception.DropboxException;
-import com.dropbox.client2.session.AccessTokenPair;
-import com.dropbox.client2.session.AppKeyPair;
-import com.dropbox.client2.session.Session.AccessType;
 
 import java.io.*;
 import java.util.ArrayList;
-import java.util.Locale;
 
 
 public class DropBoxActivity extends Activity {
 
-    final static private String APP_KEY = "2vfpyoojj4rsi5t";
-    final static private String APP_SECRET = "m6okqlqhi1nugq1";
-    final static private AccessType ACCESS_TYPE = AccessType.DROPBOX;
-
-    final static public String ACCOUNT_PREFS_NAME = "prefs";
-    final static public String ACCESS_KEY_NAME = "ACCESS_KEY";
-    final static public String ACCESS_SECRET_NAME = "ACCESS_SECRET";
-
-    public static DropboxAPI<AndroidAuthSession> mDBApi;
     ReLaunchApp app;
 
     Button droplink_btn;
@@ -56,6 +35,7 @@ public class DropBoxActivity extends Activity {
     int count_download_files = 0;
     boolean flagDelDB;
     boolean flagDelLocFile;
+    static MyDropboxClient dropboxClient;
 
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -65,23 +45,16 @@ public class DropBoxActivity extends Activity {
         }else{
             prefs = PreferenceManager.getDefaultSharedPreferences(getBaseContext());
             app = ((ReLaunchApp) getApplicationContext());
+            if(app == null ) {
+                finish();
+            }
             app.setFullScreenIfNecessary(this);
             setContentView(R.layout.dropsync);
             final Download_Dropbox[] downdrop = new Download_Dropbox[1];
             final Upload_Dropbox[] uploaddrop = new Upload_Dropbox[1];
             String str_f;
-            Locale locale;
 
             N2EpdController.n2MainActivity = this;
-
-            str_f = prefs.getString("lang", "default");
-            if (!str_f.equals("default")) {
-                locale = new Locale(str_f);
-                Locale.setDefault(locale);
-                Configuration config = new Configuration();
-                config.locale = locale;
-                getBaseContext().getResources().updateConfiguration(config, null);
-            }
 
             DBPath = prefs.getString("DropBoxfolder", "none");
             str_f = DBPath.trim();
@@ -114,20 +87,16 @@ public class DropBoxActivity extends Activity {
             str_dropbox_folder.setText(str_f + ": " + DBPath);
             str_f = getString(R.string.pref_i_Local_folder_DropBox);
             str_local_folder.setText(str_f + ": " + DBLocalPath);
-            checkAppKeySetup();
-            AndroidAuthSession session = buildSession();
-
-            mDBApi = new DropboxAPI<AndroidAuthSession>(session);
 
             droplink_btn.setOnClickListener( new View.OnClickListener() {
                 public void onClick(View v) {
-                    mDBApi.getSession().startAuthentication(DropBoxActivity.this);
+                    dropboxClient.logIn();
                 }
             } );
 
             dropunlink_btn.setOnClickListener( new View.OnClickListener() {
                 public void onClick(View v) {
-                    logOut();
+                    dropboxClient.logOut();
                     droplink_btn.setEnabled(true);
                     dropunlink_btn.setEnabled(false);
                     download_btn.setEnabled(false);
@@ -168,137 +137,52 @@ public class DropBoxActivity extends Activity {
             } );
 
             if (DBPath.equals("none") || DBLocalPath.equals("none")){
-                droplink_btn.setEnabled(false);
+                app.showToast(getString(R.string.srt_dbactivity_warn_no_config));//WARNING! No configure! Exit for configure.
+                finish();
+            }
+            dropboxClient = new MyDropboxClient(prefs, DropBoxActivity.this);
+            if (!dropboxClient.getSession()){
+                droplink_btn.setEnabled(true);
                 dropunlink_btn.setEnabled(false);
                 download_btn.setEnabled(false);
                 upload_btn.setEnabled(false);
                 dbselect_btn.setEnabled(false);
-                app.showToast(getString(R.string.srt_dbactivity_warn_no_config));//WARNING! No configure! Exit for configure.
+            }else{
+                droplink_btn.setEnabled(false);
+                dropunlink_btn.setEnabled(true);
+                download_btn.setEnabled(true);
+                upload_btn.setEnabled(true);
+                dbselect_btn.setEnabled(true);
             }
             if (flag_download == 1){
                 download_btn.setEnabled(false);
                 upload_btn.setEnabled(false);
                 dbselect_btn.setEnabled(false);
             }
-            dbselect_btn.setOnClickListener( new View.OnClickListener() {
+            dbselect_btn.setOnClickListener(new View.OnClickListener() {
                 public void onClick(View v) {
                     Intent intent = new Intent(DropBoxActivity.this, DropboxSelect.class);
                     intent.putExtra("DBLocalPath", DBLocalPath);
                     startActivity(intent);
                 }
-            } );
+            });
         }
     }
 
     protected void onResume() {
         super.onResume();
-
-        if (mDBApi.getSession().authenticationSuccessful()) {
-            try {
-                // Required to complete auth, sets the access token on the session
-                mDBApi.getSession().finishAuthentication();
-
-                AccessTokenPair tokens = mDBApi.getSession().getAccessTokenPair();
-                storeKeys(tokens.key, tokens.secret);
-
-                droplink_btn.setEnabled(false);
-                dropunlink_btn.setEnabled(true);
-                if (flag_download == 1){
-                    download_btn.setEnabled(false);
-                    upload_btn.setEnabled(false);
-                    dbselect_btn.setEnabled(false);
-                }else{
-                    download_btn.setEnabled(true);
-                    upload_btn.setEnabled(true);
-                    dbselect_btn.setEnabled(true);
-                }
-            } catch (IllegalStateException e) {
-                app.showToast(getString(R.string.srt_dbactivity_err_authent));//"Error connect for Authentication");
-                finish();
-            }
-        }
-    }
-
-    private void storeKeys(String key, String secret) {
-        // Save the access key for later
-        prefs = getSharedPreferences(ACCOUNT_PREFS_NAME, 0);
-        SharedPreferences.Editor edit = prefs.edit();
-        edit.putString(ACCESS_KEY_NAME, key);
-        edit.putString(ACCESS_SECRET_NAME, secret);
-        edit.commit();
-    }
-    private String[] getKeys() {
-        prefs = getSharedPreferences(ACCOUNT_PREFS_NAME, 0);
-        String key = prefs.getString(ACCESS_KEY_NAME, null);
-        String secret = prefs.getString(ACCESS_SECRET_NAME, null);
-        if (key != null && secret != null) {
-            String[] ret = new String[2];
-            ret[0] = key;
-            ret[1] = secret;
-            return ret;
-        } else {
-            return null;
-        }
-    }
-    private void clearKeys() {
-        prefs = getSharedPreferences(ACCOUNT_PREFS_NAME, 0);
-        SharedPreferences.Editor edit = prefs.edit();
-        edit.clear();
-        edit.commit();
-    }
-    public AndroidAuthSession buildSession() {
-        AppKeyPair appKeyPair = new AppKeyPair(APP_KEY, APP_SECRET);
-        AndroidAuthSession session;
-
-        String[] stored = getKeys();
-        if (stored != null) {
+        if (dropboxClient.logFinish()){
             droplink_btn.setEnabled(false);
             dropunlink_btn.setEnabled(true);
-            if (flag_download == 0){
+            if (flag_download == 1){
+                download_btn.setEnabled(false);
+                upload_btn.setEnabled(false);
+                dbselect_btn.setEnabled(false);
+            }else{
                 download_btn.setEnabled(true);
                 upload_btn.setEnabled(true);
                 dbselect_btn.setEnabled(true);
             }
-            AccessTokenPair accessToken = new AccessTokenPair(stored[0], stored[1]);
-            session = new AndroidAuthSession(appKeyPair, ACCESS_TYPE, accessToken);
-        } else {
-            droplink_btn.setEnabled(true);
-            dropunlink_btn.setEnabled(false);
-            download_btn.setEnabled(false);
-            upload_btn.setEnabled(false);
-            dbselect_btn.setEnabled(false);
-            session = new AndroidAuthSession(appKeyPair, ACCESS_TYPE);
-        }
-
-        return session;
-    }
-    private void logOut() {
-        // Remove credentials from the session
-        mDBApi.getSession().unlink();
-        // Clear our stored keys
-        clearKeys();
-    }
-    public void checkAppKeySetup() {
-        // Check to make sure that we have a valid app key
-        if (APP_KEY.startsWith("CHANGE") ||
-                APP_SECRET.startsWith("CHANGE")) {
-            app.showToast("You must apply for an app key and secret from developers.dropbox.com, and add them to the DBRoulette ap before trying it.");
-            finish();
-            return;
-        }
-
-        // Check if the app has set up its manifest properly.
-        Intent testIntent = new Intent(Intent.ACTION_VIEW);
-        String scheme = "db-" + APP_KEY;
-        String uri = scheme + "://" + AuthActivity.AUTH_VERSION + "/test";
-        testIntent.setData(Uri.parse(uri));
-        PackageManager pm = getPackageManager();
-        if (0 == pm.queryIntentActivities(testIntent, 0).size()) {
-            app.showToast("URL scheme in your app's " +
-                    "manifest is not set up correctly. You should have a " +
-                    "com.dropbox.client2.android.AuthActivity with the " +
-                    "scheme: " + scheme);
-            finish();
         }
     }
 
@@ -313,7 +197,6 @@ public class DropBoxActivity extends Activity {
             count_download_files = 0;
             if(!loadFiles(DBPath + File.separator, fullListFiles, DBPath.length()))
                 flagErrorDownload = true;
-            Log.i("Relaunch", "flagDelLocFile: "+ flagDelLocFile);
             if(flagDelLocFile && !flagErrorDownload){
 
                 for (String fullListFile : fullListFiles) {
@@ -354,9 +237,8 @@ public class DropBoxActivity extends Activity {
         FileOutputStream mFos;
         File file;
 
-        try {
-            entries = mDBApi.metadata(DropBox_Path, 0, null, true, null);
-        } catch (DropboxException e) {
+        entries = dropboxClient.metadata(DropBox_Path);
+        if (entries == null){
             return false;
         }
         String FullPath;
@@ -383,9 +265,7 @@ public class DropBoxActivity extends Activity {
                         } catch (FileNotFoundException e1) {
                             return false;
                         }
-                        try {
-                            mDBApi.getFile(e.path, null, mFos, null);
-                        } catch (DropboxException e1) {
+                        if (!dropboxClient.getFile(e.path, mFos)){
                             file.delete();
                             return false;
                         }
@@ -429,9 +309,7 @@ public class DropBoxActivity extends Activity {
                 flagErrorDownload = true;
             if(flagDelDB && !flagErrorDownload){
                 for (String aFullListFilesDB : fullListFilesDB) {
-                    try {
-                        mDBApi.delete(DBPath + aFullListFilesDB);
-                    } catch (DropboxException e) {
+                    if (!dropboxClient.delete(DBPath + aFullListFilesDB)) {
                         flagErrorDownload = true;
                         break;
                     }
@@ -479,9 +357,7 @@ public class DropBoxActivity extends Activity {
                     } catch (FileNotFoundException e) {
                         return false;
                     }
-                    try {
-                        mDBApi.putFile(DBPath + "/" + strPath + File.separator + f1.getName(), mFos, f1.length(), null, null);
-                    } catch (DropboxException e) {
+                    if (!dropboxClient.putFile(DBPath + "/" + strPath + File.separator + f1.getName(), mFos, f1.length())){
                         return false;
                     }
                     count_download_files++;
@@ -499,9 +375,9 @@ public class DropBoxActivity extends Activity {
     private boolean listFilesDB(String PathDB, ArrayList<String> fullListFilesDB, int start_pos_path_db){
         String strPath = PathDB.substring(start_pos_path_db);
         DropboxAPI.Entry entries;
-        try {
-            entries = mDBApi.metadata(PathDB, 0, null, true, null);
-        } catch (DropboxException e) {
+
+        entries = dropboxClient.metadata(PathDB);
+        if (entries == null) {
             return false;
         }
